fix(storage): isolate task ledger authority
