+++ packages/storage/package.json
@@ -11,7 +11,7 @@
     "./execution-stores": "./dist/execution-stores.js",
     "./root-authority": "./dist/root-authority.js",
     "./runtime-policy-stores": "./dist/runtime-policy-stores.js",
-    "./task-ledger-store": "./dist/task-ledger-store.js",
+    "./task-ledger-authority": "./dist/task-ledger-authority.js",
     "./workspace-identity": "./dist/workspace-identity.js",
     "./write-queue": "./dist/write-queue.js"
   },
