+++ packages/storage/src/task-ledger-store.ts
@@ -34,77 +34,12 @@ import {
 } from '@maka/core/task-ledger';
 import { chainWrite } from './write-queue.js';
 import { assertSafeSessionId } from './session-store.js';
-import {
-  assertStorageRootLease,
-  runWithStorageRootLease,
-  StorageRootAuthorityError,
-  type StorageRootLease,
-} from './root-authority.js';
+import { registerTaskLedgerCanonicalReader } from './task-ledger-store-internal.js';
 
 export type { TaskLedgerStore } from '@maka/core/task-ledger';
 
-const writerBrand: unique symbol = Symbol('InteractiveTaskLedgerWriter');
-const writers = new WeakSet<object>();
-const writerByLease = new WeakMap<object, InteractiveTaskLedgerWriter>();
-const writerOpeningByLease = new WeakMap<object, Promise<InteractiveTaskLedgerWriter>>();
-
-export interface TaskLedgerCanonicalReader {
-  list(sessionId: string, options?: TaskLedgerListOptions): Promise<Task[]>;
-  get(sessionId: string, id: string, options?: TaskLedgerListOptions): Promise<Task | undefined>;
-}
-
-export interface InteractiveTaskLedgerWriter extends TaskLedgerStore, TaskLedgerCanonicalReader {
-  readonly kind: 'interactive';
-  readonly access: 'write';
-  readonly [writerBrand]: true;
-}
-
 export function createTaskLedgerStore(workspaceRoot: string): TaskLedgerStore {
   return new FileTaskLedgerStore(workspaceRoot);
-}
-
-export function authenticateInteractiveTaskLedgerWriter(
-  writer: InteractiveTaskLedgerWriter,
-): InteractiveTaskLedgerWriter {
-  if (!writers.has(writer)) {
-    throw new StorageRootAuthorityError(
-      'invalid_lease',
-      'Expected an authentic interactive task ledger writer',
-    );
-  }
-  return writer;
-}
-
-export async function openInteractiveTaskLedgerStoreForWrite(
-  lease: StorageRootLease<'interactive', 'write'>,
-): Promise<InteractiveTaskLedgerWriter> {
-  await assertStorageRootLease(lease, 'interactive', 'write');
-  const existing = writerByLease.get(lease);
-  if (existing) return existing;
-  const opening = writerOpeningByLease.get(lease);
-  if (opening) return opening;
-
-  const pending = Promise.resolve().then(async () => {
-    const store = await runWithStorageRootLease(
-      lease,
-      'interactive',
-      'write',
-      async (root) => new FileTaskLedgerStore(root),
-    );
-    await assertStorageRootLease(lease, 'interactive', 'write');
-    const recoveredExisting = writerByLease.get(lease);
-    if (recoveredExisting) return recoveredExisting;
-    const writer = createInteractiveWriterFacade(lease, store);
-    writers.add(writer);
-    writerByLease.set(lease, writer);
-    return writer;
-  });
-  writerOpeningByLease.set(lease, pending);
-  try {
-    return await pending;
-  } finally {
-    if (writerOpeningByLease.get(lease) === pending) writerOpeningByLease.delete(lease);
-  }
 }
 
 class FileTaskLedgerStore implements TaskLedgerStore {
@@ -114,6 +49,10 @@ class FileTaskLedgerStore implements TaskLedgerStore {
 
   constructor(workspaceRoot: string) {
     this.sessionsRoot = join(workspaceRoot, 'sessions');
+    registerTaskLedgerCanonicalReader(this, {
+      list: (sessionId, options) => this.#listCanonical(sessionId, options),
+      get: (sessionId, id, options) => this.#getCanonical(sessionId, id, options),
+    });
   }
 
   async list(sessionId: string, options: TaskLedgerListOptions = {}): Promise<Task[]> {
@@ -133,13 +72,13 @@ class FileTaskLedgerStore implements TaskLedgerStore {
     return findTaskByRef(tasks, id);
   }
 
-  async listCanonical(sessionId: string, options: TaskLedgerListOptions = {}): Promise<Task[]> {
+  async #listCanonical(sessionId: string, options: TaskLedgerListOptions = {}): Promise<Task[]> {
     assertSafeSessionId(sessionId);
     const { tasks } = await this.readForMutateWithSource(sessionId);
     return this.applyListOptions(tasks, options);
   }
 
-  async getCanonical(
+  async #getCanonical(
     sessionId: string,
     id: string,
     options: TaskLedgerListOptions = {},
@@ -147,7 +86,7 @@ class FileTaskLedgerStore implements TaskLedgerStore {
     assertSafeSessionId(sessionId);
     if (!isSafeTaskId(id))
       throw new Error('Task id must be a stable token (alphanumeric plus . _ : -, max 64 chars)');
-    return findTaskByRef(await this.listCanonical(sessionId, options), id);
+    return findTaskByRef(await this.#listCanonical(sessionId, options), id);
   }
 
   subscribe(listener: (event: TaskLedgerChangedEvent) => void): () => void {
@@ -734,32 +673,6 @@ class FileTaskLedgerStore implements TaskLedgerStore {
       }
     }
   }
-}
-
-function createInteractiveWriterFacade(
-  lease: StorageRootLease<'interactive', 'write'>,
-  store: FileTaskLedgerStore,
-): InteractiveTaskLedgerWriter {
-  const run = <T>(operation: () => Promise<T>) =>
-    runWithStorageRootLease(lease, 'interactive', 'write', async () => operation());
-  const writer: InteractiveTaskLedgerWriter = {
-    kind: 'interactive',
-    access: 'write',
-    [writerBrand]: true,
-    list: (sessionId, options) => run(() => store.listCanonical(sessionId, options)),
-    get: (sessionId, id, options) => run(() => store.getCanonical(sessionId, id, options)),
-    create: (sessionId, drafts, context) => run(() => store.create(sessionId, drafts, context)),
-    update: (sessionId, id, patch, context) =>
-      run(() => store.update(sessionId, id, patch, context)),
-    claim: (sessionId, id, owner, context) => run(() => store.claim(sessionId, id, owner, context)),
-    claimAvailable: (sessionId, id, owner, scope, context) =>
-      run(() => store.claimAvailable(sessionId, id, owner, scope, context)),
-    settleAgentOutcome: (sessionId, id, outcome, context) =>
-      run(() => store.settleAgentOutcome(sessionId, id, outcome, context)),
-    subscribe: (listener) => store.subscribe(listener),
-  };
-  Object.freeze(writer);
-  return writer;
 }
 
 function decodeTaskSnapshots(text: string): TaskLedgerEventTaskSnapshot[] {
